Moved event sourced part to dedicated folder
