+++ samples/from_crud_to_eventsourcing/src/eventsourced/core/commandHandling.ts
