+++ samples/from_crud_to_eventsourcing/src/eventsourced/index.ts
@@ -2,7 +2,7 @@ import { startAPI } from '#core/api';
 import {
   SubscriptionToAllWithMongoCheckpoints,
   storeCheckpointInCollection,
-} from '#core/mongoDB';
+} from '#eventsourced/core/subscriptions';
 import { router } from './shoppingCarts/routes';
 import { projectToShoppingCartItem } from './shoppingCarts/shoppingCartDetails';
