+++ samples/from_crud_to_eventsourcing/src/eventsourced/shoppingCarts/productItem.ts
