+++ samples/from_crud_to_eventsourcing/tsconfig.json
@@ -49,6 +49,7 @@
     "baseUrl": "./" /* Base directory to resolve non-absolute module names. */,
     "paths": {
       "#core/*": ["./src/core/*"],
+      "#eventsourced/core/*": ["./src/eventsourced/core/*"],
       "#config": ["./config.ts"],
       "#testing/*": ["./src/testing/*"]
     } /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */,
